Update: convert all paged_attention examples from float16 to bfloat16
Convert all four paged_attention example variants to use bfloat16:
- a2a3/tensormap_and_ringbuffer/paged_attention
- a5/tensormap_and_ringbuffer/paged_attention
- a2a3/host_build_graph/paged_attention
- a5/host_build_graph/paged_attention

Changes across all variants:
- Change half/float16 to bfloat16_t/bfloat16 in kernel files
- Update golden.py dtype from "float16" to "bfloat16"
- Rename f16/fp16 variables and comments to bf16
- Align pto::Stride to Stride (matching reference style)

+++ examples/a2a3/host_build_graph/paged_attention/golden.py
@@ -6,7 +6,7 @@
 # INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT, MERCHANTABILITY, OR FITNESS FOR A PARTICULAR PURPOSE.
 # See LICENSE in the root of the software repository for the full text of the License.
 # -----------------------------------------------------------------------------------------------------------
-"""Paged Attention Golden - host_build_graph example (small scale, float16).
+"""Paged Attention Golden - host_build_graph example (small scale, bfloat16).
 
 Args layout: [query, key_cache, value_cache, block_table, context_lens, out, scale]
   - Tensors retain original multi-dimensional shapes (ContinuousTensor metadata carries shape/dtype)
@@ -33,7 +33,7 @@ ALL_CASES = {
         "block_size": 16,
         "context_len": 16,
         "max_model_len": 256,
-        "dtype": "float16",
+        "dtype": "bfloat16",
     },
     "Case2": {
         "batch": 1,
@@ -43,7 +43,7 @@ ALL_CASES = {
         "block_size": 16,
         "context_len": 64,
         "max_model_len": 256,
-        "dtype": "float16",
+        "dtype": "bfloat16",
     },
 }
